Move checkFile function to common
