+++ src/commands/sourcemaps/common/validation.ts
@@ -1,0 +1,18 @@
+import fs from 'fs'
+
+export const checkFile: (path: string) => {empty: boolean; exists: boolean} = (path: string) => {
+  try {
+    const stats = fs.statSync(path)
+    if (stats.size === 0) {
+      return {exists: true, empty: true}
+    }
+  } catch (error) {
+    if (error.code === 'ENOENT') {
+      return {exists: false, empty: false}
+    }
+    // Other kind of error
+    throw error
+  }
+
+  return {exists: true, empty: false}
+}
